Rename ArrayShortCircuiting.LastUse -> Analysis.LastUse
The analysis is no longer specific to the short-circuiting optimization.
