+++ futhark.cabal
@@ -166,6 +166,7 @@ library
       Futhark.Analysis.HORep.MapNest
       Futhark.Analysis.HORep.SOAC
       Futhark.Analysis.Interference
+      Futhark.Analysis.LastUse
       Futhark.Analysis.MemAlias
       Futhark.Analysis.Metrics
       Futhark.Analysis.Metrics.Type
@@ -349,7 +350,6 @@ library
       Futhark.Optimise.ArrayShortCircuiting
       Futhark.Optimise.ArrayShortCircuiting.ArrayCoalescing
       Futhark.Optimise.ArrayShortCircuiting.DataStructs
-      Futhark.Optimise.ArrayShortCircuiting.LastUse
       Futhark.Optimise.ArrayShortCircuiting.MemRefAggreg
       Futhark.Optimise.ArrayShortCircuiting.TopdownAnalysis
       Futhark.Optimise.MergeGPUBodies
